Move types to their own module
Move to new module, Distribution.Gentoo.Types

Move ghcLibDir function to Distribution.Gentoo.Util
