+++ src/Distribution/Gentoo/Packages.hs
@@ -33,23 +33,10 @@ import System.Directory( doesDirectoryExist
 import System.FilePath((</>))
 import Control.Monad
 
+import Distribution.Gentoo.Types
 import Distribution.Gentoo.Util
 
 -- -----------------------------------------------------------------------------
-
--- Representation of a cat/pkgname in Gentoo.  Note that this is
--- overly simplified.
-
-type Category = String
-type Pkg = String -- Package name.
-type VerPkg = String -- Package name with version.
-type VCatPkg = (Category, VerPkg)
-type Slot = String
-
--- When we are (re-)building packages, we don't care about the
--- version, just the slot.
-data Package = Package Category Pkg (Maybe Slot)
-             deriving(Eq, Ord, Show, Read)
 
 -- Package equality, ignoring the Slot (i.e. same category and package
 -- name).
@@ -118,11 +105,6 @@ pkgDBDir = "/var/db/pkg"
 -- -----------------------------------------------------------------------------
 
 -- Parsing the CONTENTS file of installed packages.
-
--- Representation of individual lines in a CONTENTS file.
-data Content = Dir BSFilePath
-             | Obj BSFilePath
-               deriving (Eq, Show, Ord)
 
 isDir         :: Content -> Bool
 isDir (Dir _) = True
